backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
fastapi = ">=0.115.12,<0.116.0"
uvicorn = { extras = ["standard"], version = ">=0.34.2,<0.35.0" }
pymilvus = { extras = ["model"], version = ">=2.5.10,<3.0.0" }
python-dotenv = ">=1.1.0,<2.0.0"
supabase = ">=2.15.2,<3.0.0"
python-jose = "^3.4.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
